BC-577 rename room-overview page
